+++ scripts/ammunition-system/actions/reload.js
@@ -1,4 +1,5 @@
 import { handleReload } from "../../feats/crossbow-feats.js";
+import { ItemSelectDialog } from "../../utils/item-select-dialog.js";
 import { getControlledActorAndToken, getEffectFromActor, getFlag, getItem, postInChat, setEffectTarget, showWarning, Updates, useAdvancedAmmunitionSystem } from "../../utils/utils.js";
 import { getSingleWeapon, getWeapons } from "../../utils/weapon-utils.js";
 import { CONJURED_ROUND_EFFECT_ID, LOADED_EFFECT_ID, MAGAZINE_LOADED_EFFECT_ID, RELOAD_AMMUNITION_IMG } from "../constants.js";
@@ -50,12 +51,8 @@ export async function reload() {
             await postReloadToChat(token, weapon);
         } else {
             // If we have no ammunition selected, or we don't have any left in the stack, we can't reload
-            const ammo = weapon.ammunition;
+            const ammo = await getAmmunition(weapon, updates);
             if (!ammo) {
-                showWarning(`${weapon.name} has no ammunition selected.`);
-                return;
-            } else if (ammo.quantity < 1) {
-                showWarning(`Not enough ammunition to reload ${weapon.name}.`);
                 return;
             }
 
@@ -261,6 +258,88 @@ export async function reloadAll() {
 
     postInChat(actor, RELOAD_AMMUNITION_IMG, `${token.name} reloads their weapons.`, "Reload", "");
     await updates.handleUpdates();
+}
+
+async function getAmmunition(weapon, updates) {
+    const ammunition = weapon.ammunition;
+
+    if (!ammunition) {
+        return await selectNewAmmunition(
+            weapon,
+            updates,
+            `You have no ammunition compatible with ${weapon.name}.`,
+            `You have no ammunition selected for your ${weapon.name}.</p><p>Select the ammunition to load.`,
+            false
+        );
+    } else if (ammunition.quantity < 1) {
+        return await selectNewAmmunition(
+            weapon,
+            updates,
+            `Not enough ammunition to reload ${weapon.name}.`,
+            `Your selected ammunition for your ${weapon.name} is empty.</p><p>Select new ammunition to load.`,
+            true
+        );
+    } else {
+        return ammunition;
+    }
+}
+
+async function selectNewAmmunition(weapon, updates, nonAvailableMessage, selectNewMessage, defaultSetAsAmmunition) {
+    const availableAmmunition = weapon.actor.itemTypes.consumable
+        .filter(item => item.consumableType === "ammo" && !item.isStowed)
+        .filter(ammo => ammo.isAmmoFor(weapon.value))
+        .filter(ammo => ammo.quantity > 0);
+
+    if (!availableAmmunition.length) {
+        showWarning(nonAvailableMessage);
+        return null;
+    }
+
+    const ammunitionMap = new Map();
+    ammunitionMap.set(
+        "Equipped",
+        availableAmmunition.map(
+            ammo => {
+                return {
+                    id: ammo.id,
+                    name: `${ammo.name} (${ammo.quantity})`,
+                    img: ammo.img
+                };
+            }
+        )
+    );
+
+    const result = await ItemSelectDialog.getItemWithOptions(
+        "Select Ammunition",
+        selectNewMessage,
+        ammunitionMap,
+        [
+            {
+                id: "set-as-ammunition",
+                label: "Set as ammunition",
+                defaultValue: defaultSetAsAmmunition
+            }
+        ]
+    );
+
+    if (!result) {
+        return null;
+    }
+
+    const selectedAmmunition = availableAmmunition.find(ammunition => ammunition.id === result.item.id);
+
+    if (result.options["set-as-ammunition"]) {
+        updates.update(
+            weapon,
+            {
+                system: {
+                    selectedAmmoId: selectedAmmunition.id
+                }
+            }
+        );
+    }
+
+    return selectedAmmunition;
 }
 
 async function postReloadToChat(token, weapon, ammunitionName) {