+++ scripts/utils/item-select-dialog.js
@@ -16,9 +16,14 @@ export class ItemSelectDialog extends Dialog {
     }
 
     static async getItem(title, header, items) {
+        const result = await this.getItemWithOptions(title, header, items, []);
+        return result?.item;
+    }
+
+    static async getItemWithOptions(title, header, items, options) {    
         let content = `
-            <div class="item-buttons" style="max-width: max-content; justify-items: center; margin: auto;">
-            <p>${header}</p>
+            <div class="item-buttons" style="min-width: 200px; max-width: max-content; justify-items: center; margin: auto;">
+            <p style="max-width: 200px;">${header}</p>
         `;
 
         for (const itemCategory of items.keys()) {
@@ -36,7 +41,7 @@ export class ItemSelectDialog extends Dialog {
                         style="display: flex; align-items: center; margin: 4px auto"
                     >
                         <img src="${item.img}" style="border: 1px solid #444; height: 1.6em; margin-right: 0.5em"/>
-                        <span>${item.name}${item.reason ? ` (${item.reason})` : ""}</span>
+                        <span>${item.name}</span>
                     </button>
                 `;
             }
@@ -44,9 +49,45 @@ export class ItemSelectDialog extends Dialog {
             content += `</fieldset>`;
         }
 
-        content += `</div>`;
-        let itemId = await new this(title, content).getItemId();
-        return Array.from(items.values()).flat().find(item => item.id === itemId);
+        if (options.length) {
+            content += `
+                <fieldset style="border: 1px solid #a1a1a1; paddng: 5px;">
+                    <legend>Options</legend>
+                    <form>
+            `;
+
+            for (const option of options) {
+                content += `
+                    <div class="form-group">
+                        <input class="option-checkbox" type="checkbox" id="${option.id}" name="${option.id}" ${option.defaultValue ? "checked" : ""}>
+                        <label for="${option.id}">${option.label}</label>
+                    </div>
+                `;
+            }
+
+            content += `
+                    </form >
+                </fieldset >
+            `;
+        }
+
+        content += `</div > `;
+
+        const itemSelectDialog = new this(title, content);
+        itemSelectDialog.selectionOptions = {};
+        for (const option of options) {
+            itemSelectDialog.selectionOptions[option.id] = option.defaultValue;
+        }
+
+        let result = await itemSelectDialog.getItemId();
+        if (!result) {
+            return null;
+        }
+
+        return {
+            item: Array.from(items.values()).flat().find(item => item.id === result.itemId),
+            options: result.options
+        };
     }
 
     activateListeners(html) {
@@ -57,12 +98,23 @@ export class ItemSelectDialog extends Dialog {
                 this.close();
             }
         );
+
+        html.find(".option-checkbox").on(
+            "change",
+            (event) => this.selectionOptions[event.currentTarget.id] = event.currentTarget.checked
+        );
+
         super.activateListeners(html);
     }
 
     async close() {
         await super.close();
-        this.result?.(this.itemId);
+        this.result?.(
+            {
+                itemId: this.itemId,
+                options: this.selectionOptions
+            }
+        );
     }
 
     async getItemId() {
@@ -71,4 +123,4 @@ export class ItemSelectDialog extends Dialog {
             this.result = result;
         });
     }
-}
+};